Cleanup for warning and memory leaks.

+++ iyan3d/trunk/Iyan3D-Android/app/src/main/jni/SGEngine2/Core/Meshes/md5.h
@@ -350,7 +350,7 @@ public:
             printf( "%s can't be opened\n", filename ) ;
         else
         {
-            while( len = fread( buffer, 1, 1024, file ) )
+            while( len = (int)fread( buffer, 1, 1024, file ) )
                 Update( buffer, len ) ;
             Final();
